django_channels/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.3
django-allauth==0.49.0
django-filter==21.1
filelock==3.6.0
fonttools==4.33.3
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kiwisolver==1.4.2
matplotlib==3.5.2
msgpack==1.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.2
zope.interface==5.4.0